requirements.txt
Flask==3.1.0
Flask_Bcrypt==1.0.1
Flask_Limiter==3.8.0
Flask_Login==0.6.3
flask_talisman==1.1.0
flask_wtf==1.2.2
mysql_connector_repackaged==0.3.1
Pillow==11.0.0
WTForms==3.2.1
onnx
onnxruntime
onnxruntime-gpu
numpy
pandas
scikit-learn
scipy
tqdm
tensorflow
transformers
torch
torchvision
torchaudio
uvicorn
multilingual_clip
faiss